rolled back to Xander's 'master' branch EXCEPT FOR REFORMATTING.
This branch exists for the purpose of MetaSploit and other tests.

I got absolutely friggin' mental over the bloody 'optimized' JavaScript, etc. which is a serious problem when debugging/diagnosing/code-reviewing.
There's more reformatting, but this was picked as the point to start this as I had the energy and, maye more importantly, finally took up reading the GIT books. The 'roll-back/merge' didn't work out like I'd expected, but at least GIT is hinting it's got some very useful 'intelligence' - none of RCSs had which I've used up to today.

End of rant. Back to work.