fix for IE8 and MochaUI: FX:Morph/object issue.
How to reproduce: click on the clock (or one of the other links which pop up an item in a MochaUI window). Then hit that link again.
In FF, the window will 'shake'. In IE8 you'll get a 'object undefined' JavaScript error.
This fix takes care of the IE8 error box by NOT performing the 'shake' F/X when the FX:Morph instance doesn't seem to be properly initialized.