FEATURE: improved backup: config.inc.php and media thumbnails are now archived too!
Plus you'll get to see a warning in the backup manager when you use the lightbox module as you must backup the media files yourself -- the PHP archiving method used simply cannot cope with large data quantities, which can quickly occur when you're showing a few photo albums through the lightbox module.

Furthermore, the backup has been altered so that the config.inc.php and sqldump files are stored in the archive so that extracting a backup archive in a CCMS root directory will automatically place all the files where they should be. The config.inc.php and sqldump filess are placed in a special /media/files/ subfolder, which will be inspected by the CCMS installer/wizard when performing an upgrade/site restore instead of a fresh install (the installer patch will occur in a later git commit).