first steps towards an installer which can perform an upgrade or site restore.
When doing a 'site restore' the installer expects the backup to be created using the backup system as per commit 4bfeb367de97118067d8489f4be511cda75f370b. When restoring from an older version's archive, you lack important data, such as the original config.inc.php: without the authcode listed in there, all passwords stored in the database are lost, PERMANENTLY. With older version's backups, you also have to move the extracted data around to the correct directories; only the backup system at the above listed commit ensures the data is stored in the archive in the same directory layout as it resides on disc.

Also, commit 4bfeb367de97118067d8489f4be511cda75f370b provides an additional feature: when a backup is performed, the sqldump and config.inc.php files are copied to a special /media/files/ subdirectory, which is inspected by the installer to decide whether to perform a fresh install or an upgrade. (NOTE that this functionality is not fully operational as per this commit! Further development is pending.)