FEATURE: Added a 'translation' screen to the ADMIN screen (for the DEVELOPMENT_ENVIRONMENT=true setting only!); this is a bit rough google translation services assistance page which helps us update translation files.
This is currently a 'development env. only!' feature as it's rather rough; as the dev.env. define in config.inc.php will be set to FALSE for any (semi-)production site anyway, that setting is checked and the page made completely UNAVAILABLE under those circumstances in order to prevent security & usability issues popping up.

----

The current process is like this:

1) open the translation page. A raw-looking dump of all translation texts is shown in a large table.

2) pick a language in the google translation toolbar shown at the top of the window

3) wait until google has translated everything, then select everything in the window by clicking in the window and then hitting CTRL+A, CTRL+C.

4) scroll down to the bottom of the page (= window) where a TinyMCE editor will be available.

5) click in the editable field and hit CTRL+V. Do not bother about the tekst already shown in there. Do not bother about the look of the content appearing; it may be a bit of a mess, but everything is FINE as long as it looks like a reasonable copy of the page HTML. The clicks and keypresses mentioned in this step-by-step guide should do the trick.

6) go to the very bottom of the page, where a blue-ish 'save' button (as usual) will be visible. Don't mind that google did translate that one as well; simply hit it and wait for the pasted content to be processed...

7) ... when all is well, the page content has been munched by the translation assistance page code and produced a suitable language.inc.php file in the /media/files/ folder and another one in the /media/files/lang-babel/ folder -- which folder may have just been created by the assistance page.

The /media/files/<langcode>.inc.php file is an UPDATED copy of the /lib/languages/<langcode>.inc.php file, which has been mixed&mashed with the google tranaslation according to these rules:

--
A) when the /lib/languages/<langcode>.inc.php file exists, any translated texts in there, i.e. any texts which are NOT IDENTICAL to the equivalent texts in /lib/languages/en.inc.php, re prevalent. That means they win over everything else. The impact is that your existing 'official' translation remains intact!

B) when google has provided a translation, no matter how crappy (we expect a human translator to validate and verify this output anyway!), then we use that one, if (A) didn't provide any.

C) anything that's left is set to the text as provided by the /lib/languages/en.inc.php file.
--

These 3 rules are particularly important for the files in /media/files/lang-babel/ : the difference between those and the ones in /media/files/ is that the ones in /media/files/ are close copies of the existing files in /lib/languages/ while the ones in /media/files/lang-babel/ are closer to the 'ideal form': they contain the same data as the other one, but this time they are based on the exact layout teplate as provided by the /lib/languages/en.inc.php file.

NOTES:

- upgrading/patching the language files in /lib/languages/ should be a 3-stage process:
(1) take the equivalent file in /media/files/ and replace the /lib/languages/ one with that new one. This way, we can have the cleanest 'diff' showing us what the filtering and google-translation process did to our existing content.
(2) next, replace the new /lib/lanaguges/ one by the one in /media/files/lang-babel/ : that one will have all the entries nicely and correctly laid out, just like they are in the original /lib/languages/en.inc.php file (which, incidentally, is ASSUMED to always be 'complete', from start to finish of this entire translation process).
(3) next, have a human translator evaluate the results. NOTEW THAT THE MACHINE TRANSLATIONS ARE CLEARLY MARKED AS SUCH: THEY ARE PREFIXED WITH THE
    /* BABELFISH */
COMMENT, so you can clearly see which ones were produced by google machine translation effort.