(re-)introduction of CSS/JS minification; now done 'on-line', in the caching back-end (combine.inc.php).
A few different CSS and JS minification facilities can be selected by configuring the $optimize[] array at the top of said file (combine.inc.php). csstidy and JSmin are currently preferred, but we'll have to see which ones perform really well (csstidy, for example, wasn't exactly 'right for the job' out of the box; further testing is required to see which of the ones included here is 'the best').