fix for restores/backups using an sqldump file created on a Windows box, i.e. a sqldump file with CRLF instead of LF-only line endings + default setting in config.inc.php is now 'production environment' so anyone grabbing the code from github will, by default, start CCMS in a 'production setting'.
The latter means that the installer will not merely dump all queries it /should/ execute, but actually /do/ the work it is expected to perform. ('development environment' means you get to see a large dump of SQL commands at the end of the install run and your environment is NOT touched. That is a very useful behaviour in a development setting, but may not be exactly what one expects when grabbing the code first time around. Particularly because this bit is, ahhhh, undocumented.)