tweak so the few popup tips in the admin screen with long texts are shown in a more legible fashion (larger tip boxes).
It's a bit of a hack; the /proper/ thing to do would be to augment the mootools Tips() class and make it use different div.tip-wrap classes based on a heuristic which inspects the tip:text' length. Couldn't get this idea to work, so here's the alternative: a brutal style/sprite hack.