partial migration to blueprint CSS 1.0 (the latest version) and using their source level CSS files through Combiner (who does the flattening ~ merging and minification)
Also first round of fixes for the sprites: in IE6/IE7/IE8 and in FF as well those exhibits bugs: when used in headers which are rendered with a blockheight > 16px (e.g. the section headers in the admin screen) some pixel rows from adjacent sprites would show through. This is particularly an issue on IE<=7

Also a partial fix for the positioning woos in the installer of the 'next' button vs. cancel 'link': blueprintCSS has styling for such matters, but it wasn't used.

Since we use various page 'widths' it's no use defining a 'container' class; instead we add container-x classes where x is the number of columns: the installer uses 18, as does the login screen, while the admin screen uses 25.

Note that blueprintCSS 1.0 now defaults in source code to 24 columns instead of 25; instead of adjusting the blueprintCSS using their own ruby script, we just add the 25th column in our base.css @import-ing wrapper CSS; the whole thing is flattened by the Combiner anyway, so this is close enough and we're able to track their edits this way (despite the fact that some minor modifications had to be made in the src/*.css blueprintCSS styles)

WARNING: THIS IS GUARANTEED NOT TO WORK (YET):
it a preliminary checkin; a few screens are working again and we're currently working on fixing the content and the CSS to make all screens render properly on IE6/7/8 and FF3 (which wasn't happening before) -- further edits will be submitted in this branch until we're back at a quality level that's sufficient to permit merging back into 'master'.